requirements.txt
appdirs==1.4.4
asgiref==3.8.1
astroid==2.8.4
certifi==2020.12.5
colorama==0.4.4
distlib==0.3.1
dj-database-url==2.3.0
Django==4.2.16
django-appconf==1.0.6
django-filter==24.3
django-heroku==0.3.1
django-select2==8.2.1
filelock==3.0.12
gunicorn==23.0.0
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
packaging==24.1
Pillow==11.1.0
pipenv==2020.11.15
platformdirs==2.4.0
psycopg2==2.9.10
pylint==2.11.1
pyzbar=0.1.9
six==1.15.0
sqlparse==0.5.1
toml==0.10.2
typing_extensions==4.12.2
tzdata==2024.2
virtualenv==20.4.3
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
whitenoise==6.7.0
wrapt==1.13.3
